Refactoring (long) 0 to 0L

+++ src/main/java/com/coralblocks/coralaffinity/pointer/_104BitsPointer.java
@@ -5,7 +5,7 @@ public class _104BitsPointer extends Pointer {
 	public static final int SIZE_IN_BYTES = 13;
 
 	public _104BitsPointer() {
-		this((long) 0, (int) 0, (byte) 0);
+		this(0L, (int) 0, (byte) 0);
 	}
 	
 	public _104BitsPointer(long l, int i, byte b) {
@@ -15,7 +15,7 @@ public class _104BitsPointer extends Pointer {
 	
 	@Override
 	public void reset() {
-		set((long) 0, (int) 0, (byte) 0);
+		set(0L, (int) 0, (byte) 0);
 	}
 	
 	private void set(long l, int i, byte b) {
