+++ src/main/java/com/coralblocks/coralaffinity/pointer/_128BitsPointer.java
@@ -5,7 +5,7 @@ public class _128BitsPointer extends Pointer {
 	public static final int SIZE_IN_BYTES = 16;
 
 	public _128BitsPointer() {
-		this((long) 0, (long) 0);
+		this(0L, 0L);
 	}
 	
 	public _128BitsPointer(long l1, long l2) {
@@ -15,7 +15,7 @@ public class _128BitsPointer extends Pointer {
 
 	@Override
 	public void reset() {
-		set((long) 0, (long) 0);
+		set(0L, 0L);
 	}
 	
 	private void set(long l1, long l2) {
