+++ src/main/java/com/coralblocks/coralaffinity/pointer/_152BitsPointer.java
@@ -5,7 +5,7 @@ public class _152BitsPointer extends Pointer {
 	public static final int SIZE_IN_BYTES = 19;
 
 	public _152BitsPointer() {
-		this((long) 0, (long) 0, (short) 0, (byte) 0);
+		this(0L, 0L, (short) 0, (byte) 0);
 	}
 	
 	public _152BitsPointer(long l1, long l2, short s, byte b) {
@@ -15,7 +15,7 @@ public class _152BitsPointer extends Pointer {
 	
 	@Override
 	public void reset() {
-		set((long) 0, (long) 0, (short) 0, (byte) 0);
+		set(0L, 0L, (short) 0, (byte) 0);
 	}
 	
 	private void set(long l1, long l2, short s, byte b) {
