+++ src/main/java/com/coralblocks/coralaffinity/pointer/_160BitsPointer.java
@@ -5,7 +5,7 @@ public class _160BitsPointer extends Pointer {
 	public static final int SIZE_IN_BYTES = 20;
 
 	public _160BitsPointer() {
-		this((long) 0, (long) 0, (int) 0);
+		this(0L, 0L, (int) 0);
 	}
 	
 	public _160BitsPointer(long l1, long l2, int i) {
@@ -15,7 +15,7 @@ public class _160BitsPointer extends Pointer {
 	
 	@Override
 	public void reset() {
-		set((long) 0, (long) 0, (int) 0);
+		set(0L, 0L, (int) 0);
 	}
 	
 	private void set(long l1, long l2, int i) {
