+++ src/main/java/com/coralblocks/coralaffinity/pointer/_176BitsPointer.java
@@ -5,7 +5,7 @@ public class _176BitsPointer extends Pointer {
 	public static final int SIZE_IN_BYTES = 22;
 
 	public _176BitsPointer() {
-		this((long) 0, (long) 0, (int) 0, (short) 0);
+		this(0L, 0L, (int) 0, (short) 0);
 	}
 	
 	public _176BitsPointer(long l1, long l2, int i, short s) {
@@ -15,7 +15,7 @@ public class _176BitsPointer extends Pointer {
 	
 	@Override
 	public void reset() {
-		set((long) 0, (long) 0, (int) 0, (short) 0);
+		set(0L, 0L, (int) 0, (short) 0);
 	}
 	
 	private void set(long l1, long l2, int i, short s) {
