+++ src/main/java/com/coralblocks/coralaffinity/pointer/_208BitsPointer.java
@@ -5,7 +5,7 @@ public class _208BitsPointer extends Pointer {
 	public static final int SIZE_IN_BYTES = 26;
 
 	public _208BitsPointer() {
-		this((long) 0, (long) 0, (long) 0, (short) 0);
+		this(0L, 0L, 0L, (short) 0);
 	}
 	
 	public _208BitsPointer(long l1, long l2, long l3, short s) {
@@ -15,7 +15,7 @@ public class _208BitsPointer extends Pointer {
 	
 	@Override
 	public void reset() {
-		set((long) 0, (long) 0, (long) 0, (short) 0);
+		set(0L, 0L, 0L, (short) 0);
 	}
 	
 	private void set(long l1, long l2, long l3, short s) {
