+++ src/main/java/com/coralblocks/coralaffinity/pointer/_240BitsPointer.java
@@ -5,7 +5,7 @@ public class _240BitsPointer extends Pointer {
 	public static final int SIZE_IN_BYTES = 30;
 
 	public _240BitsPointer() {
-		this((long) 0, (long) 0, (long) 0, (int) 0, (short) 0);
+		this(0L, 0L, 0L, (int) 0, (short) 0);
 	}
 	
 	public _240BitsPointer(long l1, long l2, long l3, int i, short s) {
@@ -15,7 +15,7 @@ public class _240BitsPointer extends Pointer {
 	
 	@Override
 	public void reset() {
-		set((long) 0, (long) 0, (long) 0, (int) 0, (short) 0);
+		set(0L, 0L, 0L, (int) 0, (short) 0);
 	}
 	
 	private void set(long l1, long l2, long l3, int i, short s) {
